Behold the magic of casting!
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@71146 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/rtp.c
@@ -1062,7 +1062,7 @@ struct ast_frame *ast_rtcp_read(struct ast_rtp *rtp)
 								    rtp->rtcp->reported_jitter,
 								    (unsigned long) ntohl(rtcpheader[i + 4]) >> 16, ((unsigned long) ntohl(rtcpheader[i + 4]) << 16) * 4096,
 								    ntohl(rtcpheader[i + 5])/65536.0,
-								    rtt);
+								    (unsigned long long)rtt);
 			} else {
 				manager_event(EVENT_FLAG_REPORTING, "RTCPReceived", "From %s:%d\r\n"
 								    "PT: %d(%s)\r\n"
